+++ com.coregame.sharedmeta/Runtime/Client/Network/DispatcherNetworkAdapter.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Collections.Immutable;
 using System.Threading.Tasks;
 using SharedMeta.Core;
 using SharedMeta.Core.Network;